+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/cli.py
@@ -1,4 +1,28 @@
-"""Console script for {{cookiecutter.project_slug}}."""
+{%- if cookiecutter.command_line_interface|lower == 'docopt' -%}
+"""
+Console script for {{cookiecutter.project_slug}}.
+Example ussage of docopt https://github.com/docopt/docopt
+
+Usage:
+  cli ship new <name>...
+  cli ship <name> move <x> <y> [--speed=<kn>]
+  cli ship shoot <x> <y>
+  cli mine (set|remove) <x> <y> [--moored|--drifting]
+  cli -h | --help
+  cli --version
+
+Options:
+  -h --help     Show this screen.
+  --version     Show version.
+  --speed=<kn>  Speed in knots [default: 10].
+  --moored      Moored (anchored) mine.
+  --drifting    Drifting mine.
+"""
+{%- else -%}
+"""
+Console script for {{cookiecutter.project_slug}}.
+"""
+{%- endif %}
 
 {%- if cookiecutter.command_line_interface|lower == 'argparse' %}
 import argparse
@@ -6,6 +30,10 @@ import argparse
 import sys
 {%- if cookiecutter.command_line_interface|lower == 'click' %}
 import click
+{%- endif %}
+{%- if cookiecutter.command_line_interface|lower == 'docopt' %}
+from docopt import docopt
+import pkg_resources
 {%- endif %}
 
 {% if cookiecutter.command_line_interface|lower == 'click' %}
@@ -27,6 +55,13 @@ def main():
     print("Arguments: " + str(args._))
     print("Replace this message by putting your code into "
           "{{cookiecutter.project_slug}}.cli.main")
+    return 0
+{%- endif %}
+{%- if cookiecutter.command_line_interface|lower == 'docopt' %}
+def main():
+    version = pkg_resources.get_distribution('{{cookiecutter.project_slug}}').version
+    arguments = docopt(__doc__, version=f'Naval Fate {version}')
+    print(arguments)
     return 0
 {%- endif %}